requirements.txt
APScheduler==3.10.1
backports.zoneinfo==0.2.1
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
httplib2==0.22.0
idna==3.4
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pytz==2023.3
PyYAML==6.0
requests==2.30.0
rsa==4.9
six==1.16.0
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.15
